Sample/requirements.txt
click==8.0.4
colorama==0.4.4
dominate==2.6.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-MySQLdb==1.0.1
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
mysqlclient==2.1.0
PyYAML==6.0
visitor==0.1.3
Werkzeug==2.0.3
